Fixed thread safety issue. Added/fixed comments
