+++ sdks/cpp/connections/REST/include/DeviceRequest.h
@@ -46,6 +46,7 @@
 
 // Connections/REST
 #include "api.h"
+#include "SocketReader.h"
 #include "SocketWriter.h"
 using catena::API;
